requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
langchain>=0.1.0
langchain-experimental>=0.0.49
langchain-groq>=0.1.0
crewai>=0.11.0
matplotlib>=3.7.0
seaborn>=0.12.0
python-dotenv>=1.0.0
markdown>=3.5.0
pydantic>=2.0.0
statsmodels==0.14.1
tenacity==8.5.0
